HomeWork6 CaseOfBanknotesImp moved to package
