+++ hw06-solid/src/main/java/ru/otus/CaseOfBanknotesImp.java
@@ -1,0 +1,48 @@
+package ru.otus;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
+public class CaseOfBanknotesImp implements CaseOfBanknotes {
+    private final Banknotes banknotes;
+    private int numbersBanknotes;
+
+    public CaseOfBanknotesImp(Banknotes banknotes, int numbersBanknotes) {
+        this.banknotes = banknotes;
+        this.numbersBanknotes = numbersBanknotes;
+    }
+
+    @Override
+    public List<Banknotes> giveBankNotes(Integer money) {
+        List<Banknotes> banknotesOut = new ArrayList<>();
+        if (numbersBanknotes > 0 && money > 0) {
+            int needNumbersOfBanknotes = money / banknotes.getNominal();
+            int numberOfBanknotestoGive = Math.min(needNumbersOfBanknotes, numbersBanknotes);
+            {
+                banknotesOut = IntStream.range(0, numberOfBanknotestoGive).mapToObj(i -> banknotes).collect(Collectors.toList());
+                numbersBanknotes -= numberOfBanknotestoGive;
+            }
+        }
+        return banknotesOut;
+    }
+
+    @Override
+    public Banknotes getTypeBanknotes() {
+        return banknotes;
+    }
+
+    @Override
+    public Integer getSumOfMoney() {
+        return numbersBanknotes * banknotes.getNominal();
+    }
+
+    @Override
+    public Integer putBankNote(Banknotes banknote) {
+        if (banknote != banknotes) {
+            throw new AtmException("Illegal BankNote");
+        }
+        return ++numbersBanknotes;
+    }
+}
